+++ app/Services/SinkHandler.php
@@ -172,6 +172,7 @@ class SinkHandler
         $chunk->{$stage . '_status'} = 'in_progress';
         $chunk->{$stage . '_message'} = null;
         $chunk->{$stage . '_size'} = null;
+        $chunk->{$stage . '_version'} = null;
         $chunk->{$stage . 'ed_at'} = null;
         $chunk->save();
         try {
@@ -186,6 +187,7 @@ class SinkHandler
         $chunk->{$stage . '_status'} = $finalState;
         if ($finalState === 'finished') {
             $chunk->{$stage . '_size'} = $result;
+            $chunk->{$stage . '_version'} = $stage === 'fetch' ? $this->src->getChunkVersion($chunk->chunk_id) : $chunk->fetch_version;
             $chunk->{$stage . 'ed_at'} = now();
         }
         $chunk->save();